Agrego vista para buscar un usuario
