+++ app/src/main/java/com/example/chotuve_android_client/MainActivity.kt
@@ -21,7 +21,7 @@ class MainActivity : AppCompatActivity() {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications,
+                R.id.navigation_home, R.id.navigation_search, R.id.navigation_notifications,
                 R.id.navigation_upload_video))
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
