+++ app/src/main/java/com/example/chotuve_android_client/ui/search/SearchViewModel.kt
@@ -1,0 +1,19 @@
+package com.example.chotuve_android_client.ui.search
+
+import android.text.Editable
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SearchViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = ""
+    }
+    val text: LiveData<String> = _text
+
+    fun updateText(this_user : Editable?) {
+        _text.value = this_user.toString()
+    }
+
+}
