+++ app/src/main/res/values/colors.xml
@@ -5,9 +5,12 @@
     <color name="colorPrimaryDark">#3700B3</color>
     <color name="colorAccent">#03DAC5</color>
 
+    <color name="white">#FFFFFF</color>
+
 
     <!-- https://material.io/resources/color/#!/?view.left=0&view.right=0 -->
     <color name="black_overlay">#66000000</color>
+    <color name="extremely_light_red">#ffebee</color>
     <color name="light_red">#f05545</color>
     <color name="primary_red">#b71c1c</color>
     <color name="dark_red">#7f0000</color>
